requirements.txt
asciichartpy==1.5.25
certifi==2025.11.12
charset-normalizer==3.4.4
crowdstrike-falconpy==1.5.4
idna==3.11
pyfiglet==1.0.4
requests==2.32.5
setuptools==80.9.0
tabulate==0.9.0
urllib3==2.5.0
